feat: reorganise types
