+++ bindings/nodejs/lib/MessageHandler.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { sendMessageAsync, messageHandlerNew, listen } from './bindings';
-import type { EventType, AccountManagerOptions, __SendMessagePayload__ } from './types';
+import type { EventType, AccountManagerOptions, __SendMessagePayload__ } from '../types';
 
 // The MessageHandler class interacts with messages with the rust bindings.
 export class MessageHandler {
