+++ bindings/nodejs/lib/index.ts
@@ -4,7 +4,7 @@
 export * from './AccountManager';
 export * from './MessageHandler';
 export * from './Account';
-export * from './types';
+export * from '../types';
 
 // @ts-ignore
 const addon = require('../build/Release/index.node');
