+++ bindings/nodejs/types/bridge.ts
@@ -1,264 +1,19 @@
 
-export enum Network {
-    Mainnet,
-    Testnet,
-}
-
-export enum OutputsToCollect {
-    None = 'None',
-    MicroTransactions = 'MicroTransactions',
-    NativeTokens = 'NativeTokens',
-    Nfts = 'Nfts',
-    All = 'All',
-}
-
-// TODO: Update Transaction interface
-export interface Transaction {
-    // TODO: Change to camelCase
-    message_id: string;
-    // TODO: Add other inclusion states
-    inclusion_state: 'Confirmed';
-    timestamp: number;
-    // TODO: Change to camelCase
-    network_id: number;
-    incoming: boolean;
-}
-
-export interface OutputResponse {
-    messageId: string;
-    transactionId: string;
-    outputIndex: number;
-    isSpent: boolean;
-    milestoneIndexBooked: number;
-    milestoneTimestampBooked: number;
-    ledgerIndex: number;
-    // TODO: Replace with proper type
-    output: any;
-}
-
-export interface OutputData {
-    outputId: string;
-    outputResponse: OutputResponse;
-    // TODO: Replace with proper type
-    output: any;
-    amount: number;
-    isSpent: boolean;
-    address: {
-        // TODO: Also add other types
-        type: 'Ed25519',
-        data: string;
-    },
-    networkId: number;
-    remainder: boolean;
-    // TODO: Replace with proper type
-    chain: any;
-}
-
-export type RemainderValueStrategy = {
-    ChangeAddress: {
-        strategy: 'ChangeAddress';
-        value: null
-    },
-    ReuseAddress: {
-        strategy: 'ReuseAddress';
-        value: null
-    },
-    AccountAddress: {
-        strategy: 'AccountAddress';
-        value: string;
-    },
-}
-
-export interface NativeTokenOptions {
-    // TODO: Change to camelCase
-    // TOOD: Should this be just "address"?
-    account_address: string;
-    // TOOD: Change to proper type
-    token_tag: any;
-    // TOOD: Change to camelCase
-    circulating_supply: number;
-    // TOOD: Change to camelCase
-    maximum_supply: number;
-}
-
-export interface NftOptions {
-    address: string;
-    // TOOD: Change to proper type
-    immutable_metadata: any;
-    // TOOD: Change to proper type
-    metadata: any;
-}
-
-// TODO: Transfer & TransferOptions should probably be merged
-export interface TransferOptions {
-    // TODO: Change to camelCase
-    remainder_value_strategy: RemainderValueStrategy;
-    // TODO: Change to proper type
-    tagged_data_payload: any;
-    // TODO: Change to camelCase
-    skip_sync: boolean;
-    // TODO: Change to camelCase
-    custom_inputs: string[];
-}
-
-export interface AccountBalance {
-    total: number;
-    available: number;
-    incoming: number;
-    outgoing: number;
-}
-
-export interface NodeInfo {
-    name: string;
-    version: string;
-    isHealthy: boolean;
-    networkId: string;
-    bech32HRP: string;
-    minPoWScore: number;
-    messagesPerSecond: number;
-    referencedMessagesPerSecond: number;
-    referencedRate: number;
-    latestMilestoneTimestamp: number;
-    latestMilestoneIndex: number;
-    confirmedMilestoneIndex: number;
-    pruningIndex: number;
-    features: string[];
-}
-
-export interface AccountSyncOptions {
-    addressIndex?: number;
-    gapLimit?: number;
-}
-
-export type Address = {
-    address: string;
-    keyIndex: number;
-    internal: boolean;
-    used: boolean;
-}
-
-// TODO: Change snakecase to camelCase
-export interface AccountMeta {
-    index: number;
-    // TODO: Should this be an enum?
-    coin_type: number;
-    alias: string;
-    public_addresses: Address[];
-    internal_addresses: Address[];
-    addresses_with_balance: Address[];
-    // TODO: Define type for outputs
-    outputs: any;
-    locked_outputs: any;
-    unspent_outputs: any;
-    // TODO: Define type for outputs
-    transactions: any;
-    pending_transactions: any;
-    account_options: {
-        output_consolidation_threshold: number;
-        automatic_output_consolidation: boolean;
-    }
-}
-
-export type EventType =
-    | '*'
-    | 'ErrorThrown'
-    | 'BalanceChange'
-    | 'NewTransaction'
-    | 'ConfirmationStateChange'
-    | 'Reattachment'
-    | 'Broadcast'
-    | 'TransferProgress';
-
-export type Auth = {
-    jwt?: string;
-    username?: string;
-    password?: string;
-}
-
-export interface AddressWithAmount {
-    address: string;
-    amount: number;
-}
-
-export interface AddressMicroAmount {
-    address: string;
-    amount: number;
-    // TODO: Change to camelCase
-    return_address: string;
-    expiration: number
-}
-
-export interface AddressNativeTokens {
-    address: string;
-    // TODO: Change to camelCase
-    native_tokens: string[];
-    // TODO: Change to camelCase
-    return_address: string;
-    expiration: number;
-}
-
-export interface AddressNftId {
-    address: string;
-    // TODO: Change to camelCase
-    nft_id: string;
-}
-
-export interface MqttBrokerOptions {
-    automaticDisconnect?: boolean;
-    // timeout in seconds
-    timeout?: number;
-    useWs?: boolean;
-    port?: number;
-    maxReconnectionAttempts?: number;
-}
-
-export type Node = {
-    url: string;
-    auth?: Auth;
-    disabled?: boolean;
-}
-
-export interface ClientOptions {
-    primaryNode?: string | Node;
-    primaryPoWNode?: string | Node;
-    node?: string | Node;
-    nodes?: Array<string | Node>;
-    network?: string;
-    mqttBrokerOptions?: MqttBrokerOptions;
-    quorumSize?: number;
-    quorumThreshold?: number;
-    localPow?: boolean;
-}
-
-export interface MnemonicSigner {
-    Mnemonic: string;
-}
-
-export interface StrongholdSigner {
-    Stronghold: {
-        password: string;
-        snapshotPath: string;
-    }
-}
-export type Signer = MnemonicSigner | StrongholdSigner;
-
-export interface AccountManagerOptions {
-    storagePath?: string;
-    clientOptions: ClientOptions;
-    signer: Signer;
-}
-
-export interface CreateAccountPayload {
-    alias: string
-}
+import type { AccountSyncOptions, CreateAccountPayload } from './account';
+import type { AddressWithAmount, AddressMicroAmount, AddressNativeTokens, AddressNftId } from './address';
+import type { ClientOptions } from './network';
+import type { OutputsToCollect, OutputData } from './output';
+import type { NativeTokenOptions, TransferOptions, NftOptions } from './transfer';
 
 type __GetAccountsMessagePayload__ = {
     cmd: 'GetAccounts'
 }
+
 type __GetAccountMessagePayload__ = {
     cmd: 'GetAccount'
     payload: string
 }
+
 type __CreateAccountMessagePayload__ = {
     cmd: 'CreateAccount'
     payload: CreateAccountPayload
