+++ bindings/nodejs/types/event.ts
@@ -1,0 +1,11 @@
+
+export type EventType =
+    | '*'
+    | 'ErrorThrown'
+    | 'BalanceChange'
+    | 'NewTransaction'
+    | 'ConfirmationStateChange'
+    | 'Reattachment'
+    | 'Broadcast'
+    | 'TransferProgress';
+
