+++ bindings/nodejs/types/index.ts
@@ -1,0 +1,10 @@
+export * from './account';
+export * from './accountManager';
+export * from './address';
+export * from './bridge';
+export * from './event';
+export * from './network';
+export * from './output';
+export * from './signer';
+export * from './transaction';
+export * from './transfer';
