+++ bindings/nodejs/types/signer.ts
@@ -1,0 +1,12 @@
+export interface MnemonicSigner {
+    Mnemonic: string;
+}
+
+export interface StrongholdSigner {
+    Stronghold: {
+        password: string;
+        snapshotPath: string;
+    }
+}
+
+export type Signer = MnemonicSigner | StrongholdSigner;
